Alinha pacote backend do GeraAnuncio
