+++ backend/ads-service/src/main/java/com/marketinghub/pipelines/aiworker/geracaoanuncios/v1/texto/service/recebeResposta/GeraAnuncioTextoRespostaRequest.java
@@ -1,0 +1,7 @@
+package com.marketinghub.pipelines.aiworker.geracaoanuncios.v1.texto.service.recebeResposta;
+
+import java.time.Instant;
+import java.util.Map;
+
+/** Dados auditáveis da resposta e da saída funcional da etapa Texto. */
+public record GeraAnuncioTextoRespostaRequest(String jobId, Instant receivedAt, String status, Map<String, Object> responsePayload, Map<String, Object> structuredOutput, String error) {}
