--- backend/ads-service/src/main/java/com/marketinghub/pipelines/facebookads/geracaoanuncios/v1/imagem/service/recebeResposta/GeraAnuncioImagemRespostaRequest.java
@@ -1,7 +1,0 @@
-package com.marketinghub.pipelines.facebookads.geracaoanuncios.v1.imagem.service.recebeResposta;
-
-import java.time.Instant;
-import java.util.Map;
-
-/** Dados auditáveis da resposta e da saída funcional da etapa Imagem. */
-public record GeraAnuncioImagemRespostaRequest(String jobId, Instant receivedAt, String status, Map<String, Object> responsePayload, Map<String, Object> structuredOutput, String error) {}
