+++ docs/registro-protocolo/padrao-backend.md
@@ -41,7 +41,7 @@
 - Endpoints internos pending canônicos aplicados no padrão `/api/internal/mois/dossie/v1/<etapa>/stage-executions/pending`.
 - Contratos da operação mantidos como `record` em subpacotes de `service`.
 
-- 2026-06-26 — Aplicado ao pipeline `geracaoanuncios` v1 no backend, pacote `com.marketinghub.pipelines.facebookads.geracaoanuncios.v1`, etapas `texto` e `imagem`, com endpoints pending canônicos `/api/internal/facebookads/geracaoanuncios/v1/texto/stage-executions/pending` e `/api/internal/facebookads/geracaoanuncios/v1/imagem/stage-executions/pending` e contratos DTO como `record` em subpacotes de service.
+- 2026-06-26 — Aplicado ao pipeline `geracaoanuncios` v1 no backend, pacote `com.marketinghub.pipelines.aiworker.geracaoanuncios.v1`, etapas `texto` e `imagem`, com endpoints pending canônicos `/api/internal/aiworker/geracaoanuncios/v1/texto/stage-executions/pending` e `/api/internal/aiworker/geracaoanuncios/v1/imagem/stage-executions/pending` e contratos DTO como `record` em subpacotes de service.
 - 2026-06-26 — Aplicado ao pipeline `geraanuncio` v2 no backend, pacote `com.marketinghub.geraanuncio.v2`, etapa inicial `criativo`, com endpoint pending canônico `/api/internal/geraanuncio/v2/criativo/stage-executions/pending` e contratos DTO como `record` em subpacotes de service.
 
 ## 2026-06-26 — MOIS dossiê do produto v1
